trigger duplicate symbols for testing

+++ plugins/input/geojson/geojson_datasource.hpp
@@ -74,7 +74,7 @@ struct options_type<geojson_linear<Max,Min> >
 class geojson_datasource : public mapnik::datasource
 {
 public:
-    using box_type = mapnik::box2d<double>;
+    using box_type = mapnik::box2d<float>;
     using item_type = std::pair<box_type, std::pair<std::size_t, std::size_t> >;
     using spatial_index_type = boost::geometry::index::rtree<item_type,geojson_linear<16,4> >;
